+++ server/routes/me.js
@@ -1,6 +1,6 @@
 const router = require('express').Router();
-const { SpotifyAPI } = require('../spotify/spotifyApi.js');
-const { SpotifyParser } = require('../spotify/spotifyParser.js');
+const { SpotifyAPI } = require('../controllers/spotify/spotifyApi.js');
+const { SpotifyParser } = require('../utils/spotify.parser.js');
 
 router.get('/info', async (req, res) => {
     let userInfo = await SpotifyAPI.Get('/me', req, res);
